+++ parrot/CHANGELOG.md
@@ -9,3 +9,19 @@ This keeps note of the API changes so I don't have to stare at commit messages a
 - Included wgpu spirv feature and hence, make the function safe
 
 - Re-exported painter in the library
+
+## Verion 0.0.3
+
+### Changes
+
+- Improved documentation (wow)
+
+- New type of Set to prevent typing out verbose arrays of arrays
+
+- Added function in painter to create index buffers
+
+- Vertex and index buffers are now labled
+
+- Pipeline has been edited to allow for no buffers
+
+- Plumbers now have to know what user defined vertex type they're using